Update main
Create a separate method for checking an IP address on metadefender as well as displaying information gathered about it from both VirusTotal and MetaDefender. I also added "if" statements to the "check_ip_virustotal" method as I noticed that sometimes the information fetched doesn't include these fields causing the program to crash as a result.